Plugin Config Files + Kernel Size + FFI Fix
Implemented a mechanism to allow plugins to manage their own config files in a safe way.
Added an optional `kernel_size` parameter for the `gaussian` and `inv_gaussian` filters whose default value is provided in a plugin config file.
Fixed an FFI issue with passing the plugin trait object that caused weird (and possibly undefined) behavior.