Sound board toggle buttons
Added the option to specify a deactivate command for sound board buttons, transforming them into toggle buttons which execute the deactivate command on every second push. This resolves #136
Also, reworked the factorization of guild state and board manager accesses in the command module to use guards instead of closures (due to issues with async usage).